+++ eoreader/utils.py
@@ -374,27 +374,6 @@ def simplify(footprint_fct: Callable):
     return simplify_wrapper
 
 
-def stack_dict(
-    bands: list, band_xds: xr.Dataset, save_as_int: bool, nodata: float, **kwargs
-) -> (xr.DataArray, type):
-    """
-    Stack a dictionary containing bands in a DataArray
-
-    Args:
-        bands (list): List of bands (to keep the right order of the stack)
-        band_xds (xr.Dataset): Dataset containing the bands
-        save_as_int (bool): Convert stack to uint16 to save disk space (and therefore multiply the values by 10.000)
-        nodata (float): Nodata value
-
-    Returns:
-        (xr.DataArray, type): Stack as a DataArray and its dtype
-    """
-    logs.deprecation_warning(
-        "Deprecated function. Please use `utils.stack` instead. 'bands' is not necessary anymore"
-    )
-    return stack(band_xds, save_as_int, nodata, **kwargs)
-
-
 def stack(
     band_xds: xr.Dataset, save_as_int: bool, nodata: float, **kwargs
 ) -> (xr.DataArray, type):